+++ apps/learn-card-app/src/components/qrcode-user-card/QRCodeUserCard.tsx
@@ -7,7 +7,21 @@ import QRCodeUserCardShareOptions from './QrCodeUserCardShareOptions/QrCodeUserC
 import { useWallet, useGetCurrentLCNUser } from 'learn-card-base';
 import { getProfileIdFromLCNDidWeb } from 'learn-card-base/helpers/credentialHelpers';
 
-const QrCodeUserCard: React.FC = () => {
+const QrCodeUserCard: React.FC<{
+    showBackButton?: boolean;
+    handleBackButton?: () => void;
+    handleClose?: () => void;
+    cardTitle?: React.ReactNode | string | null;
+    contractUri?: string;
+    overrideShareLink?: string;
+}> = ({
+    showBackButton,
+    handleBackButton,
+    handleClose,
+    cardTitle = null,
+    contractUri,
+    overrideShareLink,
+}) => {
     const { initWallet } = useWallet();
     const { currentLCNUser } = useGetCurrentLCNUser();
 
@@ -43,21 +57,39 @@ const QrCodeUserCard: React.FC = () => {
     const profileId = currentLCNUser?.profileId || getProfileIdFromLCNDidWeb(walletDid);
 
     return (
-        <>
-            <div className="h-full w-full overflow-y-auto bg-white relative">
-                <QrCodeUserCardHeader showCompact={showCompactHeader} />
+        <div className="h-full w-full overflow-y-auto bg-white relative">
+            <QrCodeUserCardHeader
+                showCompact={showCompactHeader}
+                showBackButton={showBackButton}
+                handleBackButton={handleBackButton}
+                handleClose={handleClose}
+            />
 
-                <div id="qr-code-user-card-screenshot" className="flex flex-col gap-2">
-                    <div ref={basicInfoRef}>
-                        <QrCodeUserCardBasicInfo profileId={profileId} walletDid={walletDid} />
-                    </div>
+            {cardTitle && (
+                <div className="text-grayscale-900 text-lg font-semibold w-full">{cardTitle}</div>
+            )}
 
-                    <UserQRCode profileId={profileId} walletDid={walletDid} />
-                </div>
-
-                <QRCodeUserCardShareOptions />
+            <div ref={basicInfoRef}>
+                <QrCodeUserCardBasicInfo
+                    profileId={profileId}
+                    walletDid={walletDid}
+                    contractUri={contractUri}
+                    overrideShareLink={overrideShareLink}
+                />
             </div>
-        </>
+
+            <UserQRCode
+                profileId={profileId}
+                walletDid={walletDid}
+                contractUri={contractUri}
+                overrideShareLink={overrideShareLink}
+            />
+            <QRCodeUserCardShareOptions
+                contractUri={contractUri}
+                profileId={profileId}
+                overrideShareLink={overrideShareLink}
+            />
+        </div>
     );
 };